+++ library/models/sdxl/loader.py
@@ -5,7 +5,6 @@ from typing import Literal, cast
 
 from accelerate import init_empty_weights
 
-
 from library.utils.device_utils import clean_memory_on_device
 from library.utils.torch_utils import match_mixed_precision
 from library.models.runtime_utils import set_padding_mode_for_vae_conv2d_modules